STORE-2384: added svg to mime types

+++ server/openstorefront/openstorefront-core/service/src/main/java/edu/usu/sdl/openstorefront/service/component/SubComponentServiceImpl.java
@@ -191,42 +191,48 @@ public class SubComponentServiceImpl
 
 	void removeLocalResource(ComponentResource componentResource)
 	{
-		//Note: this can't be rolled back
-		ComponentResource example = new ComponentResource();
-		example.setFile(new MediaFile());
-		example.getFile().setMediaFileId(componentResource.getFile().getMediaFileId());
+		if (componentResource.getFile() != null) {
+			//Note: this can't be rolled back
+			ComponentResource example = new ComponentResource();
+			example.setFile(new MediaFile());
+			example.getFile().setMediaFileId(componentResource.getFile().getMediaFileId());
 
-		long count = persistenceService.countByExample(example);
-		if (count == 1) {
-			removeLocalMedia(componentResource.getFile().getMediaFileId());
+			long count = persistenceService.countByExample(example);
+			if (count == 1) {
+				removeLocalMedia(componentResource.getFile().getMediaFileId());
+			}
 		}
 	}
 
 	void removeLocalMedia(ComponentMedia componentMedia)
 	{
-		//Note: this can't be rolled back
-		ComponentMedia example = new ComponentMedia();
-		example.setFile(new MediaFile());
-		example.getFile().setMediaFileId(componentMedia.getFile().getMediaFileId());
+		if (componentMedia.getFile() != null) {
+			//Note: this can't be rolled back
+			ComponentMedia example = new ComponentMedia();
+			example.setFile(new MediaFile());
+			example.getFile().setMediaFileId(componentMedia.getFile().getMediaFileId());
 
-		long count = persistenceService.countByExample(example);
-		if (count == 1) {
-			removeLocalMedia(componentMedia.getFile().getMediaFileId());
+			long count = persistenceService.countByExample(example);
+			if (count == 1) {
+				removeLocalMedia(componentMedia.getFile().getMediaFileId());
+			}
 		}
 	}
 
 	void removeLocalMedia(String mediaFileId)
 	{
 		MediaFile mediaFile = persistenceService.findById(MediaFile.class, mediaFileId);
-		Path path = mediaFile.path();
-		if (path != null) {
-			if (path.toFile().exists()) {
-				if (path.toFile().delete() == false) {
-					LOG.log(Level.WARNING, MessageFormat.format("Unable to delete local media. Path: {0}", path.toString()));
+		if (mediaFile != null) {
+			Path path = mediaFile.path();
+			if (path != null) {
+				if (path.toFile().exists()) {
+					if (path.toFile().delete() == false) {
+						LOG.log(Level.WARNING, MessageFormat.format("Unable to delete local media. Path: {0}", path.toString()));
+					}
 				}
 			}
+			persistenceService.delete(mediaFile);
 		}
-		persistenceService.delete(mediaFile);
 	}
 
 	public <T extends BaseComponent> void deleteAllBaseComponent(Class<T> subComponentClass, String componentId)
